AAE-51058 Refactor query service specifications and update package structure
