+++ activiti-cloud-query-service/activiti-cloud-services-query/activiti-cloud-services-query-repo/pom.xml
@@ -15,6 +15,14 @@
       <artifactId>activiti-cloud-services-query-model</artifactId>
     </dependency>
     <dependency>
+      <groupId>org.activiti.cloud</groupId>
+      <artifactId>activiti-cloud-services-query-common</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>org.activiti.cloud</groupId>
+      <artifactId>activiti-cloud-service-common-config</artifactId>
+    </dependency>
+    <dependency>
       <groupId>org.springframework.data</groupId>
       <artifactId>spring-data-commons</artifactId>
     </dependency>
