+++ activiti-cloud-query-service/activiti-cloud-services-query/activiti-cloud-services-query-repo/src/main/java/org/activiti/cloud/services/query/app/filter/VariableType.java
@@ -13,23 +13,28 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.activiti.cloud.services.query.rest.filter;
+package org.activiti.cloud.services.query.app.filter;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.annotation.JsonValue;
 
-public enum FilterOperator {
-    @JsonProperty("eq")
-    EQUALS,
-    @JsonProperty("ne")
-    NOT_EQUALS,
-    @JsonProperty("like")
-    LIKE,
-    @JsonProperty("gt")
-    GREATER_THAN,
-    @JsonProperty("gte")
-    GREATER_THAN_OR_EQUAL,
-    @JsonProperty("lt")
-    LESS_THAN,
-    @JsonProperty("lte")
-    LESS_THAN_OR_EQUAL,
+public enum VariableType {
+    STRING,
+    INTEGER,
+    BIGDECIMAL,
+    BOOLEAN,
+    DATE,
+    DATETIME;
+
+    @JsonValue
+    public String getValue() {
+        return name().toLowerCase();
+    }
+
+    public static VariableType fromString(String name) {
+        try {
+            return VariableType.valueOf(name.toUpperCase());
+        } catch (IllegalArgumentException e) {
+            throw new IllegalArgumentException(String.format("Cannot determine variable type from '%s'", name));
+        }
+    }
 }
