+++ activiti-cloud-query-service/activiti-cloud-services-query/activiti-cloud-services-query-repo/src/main/java/org/activiti/cloud/services/query/app/specification/VariableValueFilterCondition.java
@@ -13,14 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.activiti.cloud.services.query.rest.payload;
+package org.activiti.cloud.services.query.app.specification;
 
-import java.util.Set;
-import org.activiti.cloud.services.query.rest.filter.VariableFilter;
+import jakarta.persistence.criteria.Predicate;
 
-public interface CloudRuntimeEntityFilterRequest {
-    Set<String> id();
-    Set<String> parentId();
-    Set<VariableFilter> processVariableFilters();
-    CloudRuntimeEntitySort sort();
+public interface VariableValueFilterCondition {
+    Predicate getPredicate();
 }
