+++ activiti-cloud-query-service/activiti-cloud-services-query/activiti-cloud-services-query-rest/src/main/java/org/activiti/cloud/services/query/rest/RestrictedProcessInstanceCountCacheKey.java
@@ -15,7 +15,7 @@
  */
 package org.activiti.cloud.services.query.rest;
 
-import org.activiti.cloud.services.query.rest.payload.ProcessInstanceSearchRequest;
+import org.activiti.cloud.services.query.app.payload.ProcessInstanceSearchRequest;
 
 public record RestrictedProcessInstanceCountCacheKey(
     String authenticatedUserId,
