+++ activiti-cloud-query-service/activiti-cloud-services-query/activiti-cloud-services-query-rest/src/test/java/org/activiti/cloud/services/query/util/ProcessInstanceSearchRequestBuilder.java
@@ -20,10 +20,10 @@ import java.util.Date;
 import java.util.Set;
 import java.util.stream.Collectors;
 import org.activiti.api.process.model.ProcessInstance;
+import org.activiti.cloud.services.query.app.filter.VariableFilter;
+import org.activiti.cloud.services.query.app.payload.CloudRuntimeEntitySort;
+import org.activiti.cloud.services.query.app.payload.ProcessInstanceSearchRequest;
 import org.activiti.cloud.services.query.model.ProcessVariableKey;
-import org.activiti.cloud.services.query.rest.filter.VariableFilter;
-import org.activiti.cloud.services.query.rest.payload.CloudRuntimeEntitySort;
-import org.activiti.cloud.services.query.rest.payload.ProcessInstanceSearchRequest;
 import org.springframework.data.domain.Sort;
 import tools.jackson.databind.introspect.VisibilityChecker;
 import tools.jackson.databind.json.JsonMapper;
